requirements.txt
blinker==1.8.1
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
azure-cognitiveservices-search-websearch
beautifulsoup4==4.9.3
cachetools
chardet
Flask-RESTful
htmltabletomd
lxml
lxml_html_clean
matplotlib
mplfinance
newspaper3k
openpyxl
pandas
python-docx
readabilipy
requests
tiktoken
urllib3
waitress
w3lib
pdfplumber
pytest~=6.2.5
apscheduler~=3.9.0
pytz
